refactor xcm helper
